requirements.txt
confusable-homoglyphs==3.0.0
Django==1.11.10
django-bootstrap3==9.1.0
django-registration==2.4.1
django-tinymce==2.7.0
djangorestframework==3.7.7
Pillow==5.0.0
pkg-resources==0.0.0
psycopg2==2.7.4
python-decouple==3.1
pytz==2018.3
